make WorkspaceGenerator an abstract base class
