+++ bin/carthage-play
@@ -39,7 +39,7 @@ class Command < CLAide::Command
   end
 
   def run
-    generator = Pod::WorkspaceGenerator.new([@name], :carthage)
+    generator = Pod::CarthageGenerator.new([@name])
     generator.generate
   end
 end
