+++ lib/cocoapods-playgrounds/command.rb
@@ -3,3 +3,5 @@
 require 'cocoapods-playgrounds/playground'
 require 'cocoapods-playgrounds/workspace'
 require 'cocoapods-playgrounds/command/playgrounds'
+require 'cocoapods-playgrounds/generator/carthage'
+require 'cocoapods-playgrounds/generator/cocoapods'
